feat: component 분리
